requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.7.0
anyio==4.5.0
asgiref==3.8.1
attrs==24.2.0
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
distro==1.9.0
Django==5.1.1
django-ckeditor==6.7.1
django-ckeditor-5==0.2.13
django-js-asset==2.2.0
django-modeltranslation==0.19.9
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
filelock==3.16.1
flower==2.0.1
frozenlist==1.4.1
fsspec==2024.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.25.0
humanize==4.10.0
idna==3.10
inflection==0.5.1
jiter==0.5.0
kombu==5.4.2
multidict==6.1.0
numpy==2.1.1
openai==0.28.0
packaging==24.1
pillow==10.4.0
prometheus_client==0.21.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
redis==5.0.8
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
setuptools==74.1.1
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
tokenizers==0.19.1
tornado==6.4.1
tqdm==4.66.5
transformers==4.44.2
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
yarl==1.11.1
